emailService.js created, env variables added
