+++ Server/package.json
@@ -26,6 +26,7 @@
     "mjml": "^1.3.4",
     "mongoose": "^8.3.3",
     "multer": "1.4.5-lts.1",
+    "nodemailer": "^6.9.14",
     "ping": "0.4.4",
     "sharp": "0.33.4",
     "winston": "^3.13.0"
